Move parse resolver and related objects to relevant folders
